*** empty log message ***
git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@8454 61f9eca8-113b-463c-a1a4-bab9b0a29f14

+++ src/C/WND_CHARM/cmatrix.h
@@ -90,7 +90,7 @@ class ImageMatrix
    ~ImageMatrix();                                 /* destructor */
    ImageMatrix *duplicate();                       /* create a new identical matrix        */
    void diff(ImageMatrix *matrix);                 /* compute the difference from another image */
-   void normalize(double min, double max, int range);
+   void normalize(double min, double max, long range);
    void to8bits();
    void flip();                                    /* flip an image horizonatally          */
    void Downsample(double x_ratio, double y_ratio);/* down sample an image                 */
